+++ app/src/main/java/exp/ftxt/modules/fps/FpsConfig.java
@@ -10,4 +10,9 @@ public class FpsConfig {
     public static int color = Color.WHITE;
     public static ShadowConfig shadow = new ShadowConfig();
     public static boolean touchPassthrough = false;
+    public static boolean bgEnabled = false;
+    public static int bgColor = 0xCC000000;
+    public static int bgPadding = 10;
+    public static int bgOffsetX = 0;
+    public static int bgOffsetY = 0;
 }