+++ app/src/main/java/exp/ftxt/modules/text/TextConfig.java
@@ -10,4 +10,9 @@ public class TextConfig {
     public static int color = Color.WHITE;
     public static boolean touchPassthrough = false;
     public static ShadowConfig shadow = new ShadowConfig();
+    public static boolean bgEnabled = false;
+    public static int bgColor = 0xCC000000;
+    public static int bgPadding = 25;
+    public static int bgOffsetX = 0;
+    public static int bgOffsetY = 0;
 }